requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.10.0
apache-airflow-core==3.0.2
apache-airflow-providers-apache-spark==5.3.0
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.2
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-sqlite==4.0.2
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.2
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cadwyn==5.4.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.4
debugpy==1.8.14
decorator==5.2.1
delta-spark==3.3.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.7.0
fsspec==2025.5.1
google-re2==1.1.20240702
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy-object-proxy==1.11.0
libcst==1.8.2
limits==5.3.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.2.6
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ordered-set==4.1.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
prison==0.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyspark==3.5.6
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.14.7
rpds-py==0.25.1
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tornado==6.5.1
traitlets==5.14.3
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2025.0.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.20.1
zipp==3.22.0
